Переместила правильно код,для входа
