+++ SignUpWindow.xaml
@@ -25,8 +25,8 @@
          materialDesign:HintAssist.HelperText="Helper text longer than error"
          materialDesign:HintAssist.Hint="Name" />
         <PasswordBox x:Name="Password" Grid.Column="1" VerticalAlignment="Center" Margin="0 0 35 -30" materialDesign:HintAssist.HelperText="At least 8 characters" materialDesign:HintAssist.Hint="Password" materialDesign:TextFieldAssist.HasClearButton="True" />
-        <Button Grid.Column="1" Content="Create Account" Width="Auto" Height="Auto" HorizontalAlignment="Center" VerticalAlignment="Bottom" Background="#07575B" Foreground="White" Margin="0 0 0 50" Click="Enter_Click"/>
-        <Button Grid.Column="1" Content="Already have an account?" Width="Auto" Height="Auto" HorizontalAlignment="Center" VerticalAlignment="Bottom" Margin="0 0 0 20" Click="CreateAccount_Click"/>
+        <Button Grid.Column="1" Content="Create Account" Width="Auto" Height="Auto" HorizontalAlignment="Center" VerticalAlignment="Bottom" Background="#07575B" Foreground="White" Margin="0 0 0 50" Click="CreateAccount_Click"/>
+        <Button Grid.Column="1" Content="Already have an account?" Width="Auto" Height="Auto" HorizontalAlignment="Center" VerticalAlignment="Bottom" Margin="0 0 0 20" Click="Enter_Click"/>
         <materialDesign:Snackbar x:Name="Snackbar" Canvas.Left="241" Canvas.Top="22">
             <materialDesign:SnackbarMessage Content="Успешный вход!" Foreground="White"/>
         </materialDesign:Snackbar>
